// projects/dts-backoffice-util/src/lib/interfaces/filter-range.interface.ts
export interface IFilterRangeCharacter {
  iniInitial: string;
  iniFinal: string;
  valInitial: string;
  valFinal: string;
}

export interface IFilterRangeNumber {
  iniInitial: number;
  iniFinal: number;
  valInitial: number;
  valFinal: number;
}

export interface IFilterRangeDate {
  iniInitial: Date;
  iniFinal: Date;
  valInitial: Date;
  valFinal: Date;
}